backend-py/requirements.txt
python-socketio==5.12.1
uvicorn==0.34.0
